requirements.txt
dash==3.0.4
fastapi==0.115.12
joblib==1.5.0
numpy==2.2.5
uvicorn==0.34.2
gunicorn
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
dash-bootstrap-components==2.0.3
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Flask==3.0.3
flatbuffers==25.2.10
greenlet==3.2.2
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.40.0
nest-asyncio==1.6.0
numpy==2.2.6
openmeteo_requests==1.4.0
openmeteo_sdk==1.20.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.8
plotly==6.1.1
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
retrying==1.3.4
setuptools==80.8.0
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
tornado==6.5
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
uuid==1.30
wcwidth==0.2.13
Werkzeug==3.0.6
zipp==3.21.0